split out home and senators to their own files
clears out pages/index quite a bit
